requirements.txt
streamlit
langchain-google-genai
langchain-core
pandas
python-dotenv